chore: asbtracted the memory profiler as a testing helper
